refactor(narrative): Refactor itinerary narrative rendering to allow custom renderers
Make NarrativeItinerary an abstract class to be implemented by specific renderers. Make existing
implementation "default" renderer.
