+++ lib/components/narrative/default/access-leg.js
@@ -1,9 +1,9 @@
-import Icon from './icon'
 import React, {PropTypes, Component} from 'react'
 
-import { distanceString } from '../../util/distance'
-import { getStepInstructions } from '../../util/itinerary'
-import { formatDuration } from '../../util/time'
+import Icon from '../icon'
+import { distanceString } from '../../../util/distance'
+import { getStepInstructions } from '../../../util/itinerary'
+import { formatDuration } from '../../../util/time'
 
 export default class AccessLeg extends Component {
   static propTypes = {
