+++ lib/components/narrative/default/default-itinerary.js
@@ -1,0 +1,45 @@
+import React from 'react'
+
+import NarrativeItinerary from '../narrative-itinerary'
+import ItinerarySummary from './itinerary-summary'
+import ItineraryDetails from './itinerary-details'
+import { formatDuration, formatTime } from '../../../util/time'
+
+export default class DefaultItinerary extends NarrativeItinerary {
+  render () {
+    const {
+      active,
+      activeLeg,
+      activeStep,
+      expanded,
+      index,
+      itinerary,
+      setActiveLeg,
+      setActiveStep
+    } = this.props
+    return (
+      <div className={`option itinerary${active ? ' active' : ''}`}>
+        <button
+          className='header'
+          onClick={this._onHeaderClick}
+        >
+          <span className='title'>Itinerary {index + 1}</span>{' '}
+          <span className='duration pull-right'>{formatDuration(itinerary.duration)}</span>{' '}
+          <span className='arrivalTime'>{formatTime(itinerary.startTime)}—{formatTime(itinerary.endTime)}</span>
+          <ItinerarySummary itinerary={itinerary} />
+        </button>
+        {(active || expanded) &&
+          <div className='body'>
+            <ItineraryDetails
+              itinerary={itinerary}
+              activeLeg={activeLeg}
+              activeStep={activeStep}
+              setActiveLeg={setActiveLeg}
+              setActiveStep={setActiveStep}
+            />
+          </div>
+        }
+      </div>
+    )
+  }
+}
