+++ lib/components/narrative/default/itinerary-details.js
@@ -2,7 +2,7 @@ import React, { Component, PropTypes } from 'react'
 
 import AccessLeg from './access-leg'
 import TransitLeg from './transit-leg'
-import { isTransit } from '../../util/itinerary'
+import { isTransit } from '../../../util/itinerary'
 
 export default class ItineraryDetails extends Component {
   static propTypes = {
