+++ lib/components/narrative/default/itinerary-summary.js
@@ -1,8 +1,8 @@
 import React, { Component, PropTypes } from 'react'
 import { Label } from 'react-bootstrap'
 
-import ModeIcon from './mode-icon'
-import {isTransit} from '../../util/itinerary'
+import ModeIcon from '../mode-icon'
+import {isTransit} from '../../../util/itinerary'
 
 export default class ItinerarySummary extends Component {
   static propTypes = {
