+++ lib/components/narrative/default/transit-leg.js
@@ -1,10 +1,10 @@
-import Icon from './icon'
+import Icon from '../icon'
 import React, { Component, PropTypes } from 'react'
 import { Label } from 'react-bootstrap'
 
-import ModeIcon from './mode-icon'
-import { getMapColor } from '../../util/itinerary'
-import { formatDuration } from '../../util/time'
+import ModeIcon from '../mode-icon'
+import { getMapColor } from '../../../util/itinerary'
+import { formatDuration } from '../../../util/time'
 
 export default class TransitLeg extends Component {
   static propTypes = {
