+++ lib/components/narrative/narrative-itineraries.js
@@ -2,13 +2,14 @@ import React, { Component, PropTypes } from 'react'
 import { connect } from 'react-redux'
 
 import { setActiveItinerary, setActiveLeg, setActiveStep } from '../../actions/narrative'
-import NarrativeItinerary from './narrative-itinerary'
+import DefaultItinerary from './default/default-itinerary'
 import Loading from './loading'
 import { getActiveSearch } from '../../util/state'
 
 class NarrativeItineraries extends Component {
   static propTypes = {
     itineraries: PropTypes.array,
+    itineraryClass: PropTypes.func,
     pending: PropTypes.bool,
     activeItinerary: PropTypes.number,
     setActiveItinerary: PropTypes.func,
@@ -16,8 +17,12 @@ class NarrativeItineraries extends Component {
     setActiveStep: PropTypes.func
   }
 
+  static defaultProps = {
+    itineraryClass: DefaultItinerary
+  }
+
   render () {
-    const { itineraries, pending, activeItinerary } = this.props
+    const { itineraries, pending, activeItinerary, itineraryClass } = this.props
     if (pending) return <Loading />
     if (!itineraries) return null
 
@@ -25,15 +30,13 @@ class NarrativeItineraries extends Component {
       <div className='options itinerary'>
         <div className='header'>We found {itineraries.length} itineraries:</div>
         {itineraries.map((itinerary, index) => {
-          return (
-            <NarrativeItinerary
-              itinerary={itinerary}
-              index={index}
-              key={index}
-              active={index === activeItinerary}
-              {...this.props}
-            />
-          )
+          return React.createElement(itineraryClass, {
+            itinerary,
+            index,
+            key: index,
+            active: index === activeItinerary,
+            ...this.props
+          })
         })}
       </div>
     )
