+++ lib/components/narrative/narrative-itinerary.js
@@ -1,9 +1,4 @@
-import React, { Component, PropTypes } from 'react'
-// import Icon from './icon'
-
-import ItinerarySummary from './itinerary-summary'
-import ItineraryDetails from './itinerary-details'
-import { formatDuration, formatTime } from '../../util/time'
+import { Component, PropTypes } from 'react'
 
 export default class NarrativeItinerary extends Component {
   static propTypes = {
@@ -13,6 +8,7 @@ export default class NarrativeItinerary extends Component {
     expanded: PropTypes.bool,
     index: PropTypes.number,
     itinerary: PropTypes.object,
+    onClick: PropTypes.func,
     setActiveItinerary: PropTypes.func,
     setActiveLeg: PropTypes.func,
     setActiveStep: PropTypes.func
@@ -30,39 +26,6 @@ export default class NarrativeItinerary extends Component {
   }
 
   render () {
-    const {
-      active,
-      activeLeg,
-      activeStep,
-      expanded,
-      index,
-      itinerary,
-      setActiveLeg,
-      setActiveStep
-    } = this.props
-    return (
-      <div className={`option itinerary${active ? ' active' : ''}`}>
-        <button
-          className='header'
-          onClick={this._onHeaderClick}
-        >
-          <span className='title'>Itinerary {index + 1}</span>{' '}
-          <span className='duration pull-right'>{formatDuration(itinerary.duration)}</span>{' '}
-          <span className='arrivalTime'>{formatTime(itinerary.startTime)}—{formatTime(itinerary.endTime)}</span>
-          <ItinerarySummary itinerary={itinerary} />
-        </button>
-        {(active || expanded) &&
-          <div className='body'>
-            <ItineraryDetails
-              itinerary={itinerary}
-              activeLeg={activeLeg}
-              activeStep={activeStep}
-              setActiveLeg={setActiveLeg}
-              setActiveStep={setActiveStep}
-            />
-          </div>
-        }
-      </div>
-    )
+    throw new Error('render() called on abstract class NarrativeItinerary')
   }
 }
