+++ lib/components/narrative/narrative.css
@@ -22,23 +22,23 @@
 
 /* ITINERARY OPTION */
 
-.otp .option {
+.otp .option .default-itin {
   margin-bottom: 16px;
   padding: 10px;
-  background-color: #f5f5f5;
+  background-color: green;
   border: 1px solid #e3e3e3;
   border-radius: 4px
 }
 
-.otp .option:hover {
+.otp .option:hover .default-itin {
   background-color: var(--hover-color);
 }
 
-.otp .option.active {
+.otp .option.active .default-itin {
   background-color: #EAF3F2;
 }
 
-.otp .option > .header {
+.otp .option .default-itin > .header {
   font-size: 16px;
   /*padding-bottom: 4px;*/
   /*border-bottom: 1px solid black;*/
@@ -46,43 +46,43 @@
   cursor: pointer;
 }
 
-.otp .option > .header > .title {
+.otp .option .default-itin > .header > .title {
   font-weight: bold;
 }
 
-.otp .option > .header > .summary {
+.otp .option .default-itin > .header > .summary {
   text-align: center;
   margin: 8px 0px;
 }
 
 /* ITINERARY LEG */
 
-.otp .option .leg {
+.otp .option .default-itin .leg {
   background-color: #d0d0d0;
   padding: 6px;
   margin-bottom: 8px;
 }
 
-.otp .option .leg:hover {
+.otp .option .default-itin .leg:hover {
   background-color: var(--hover-color);
 }
 
-.otp .option .leg.active {
+.otp .option .default-itin .leg.active {
   background-color: #ddd;
 }
 
-.otp .option .leg > .header {
+.otp .option .default-itin .leg > .header {
   cursor: pointer;
 }
 
-.otp .option .leg .stop-count {
+.otp .option .default-itin .leg .stop-count {
   cursor: pointer;
   /*margin-bottom: 4px;*/
   padding: 4px;
   margin-left: 8px;
 }
 
-.otp .option .leg .stop-list {
+.otp .option .default-itin .leg .stop-list {
   cursor: pointer;
   margin-bottom: 4px;
   padding: 4px;
@@ -91,37 +91,26 @@
   list-style: none;
 }
 
-.otp .option .leg .item {
+.otp .Itin .option .default-itin .leg .item {
   padding: 2px;
   font-size: small;
 }
 
-.otp .option .leg .alert-item {
+.otp .option .default-itin .leg .alert-item {
   padding-left: 22px;
 }
 
-.otp .option .leg .stop-list .stop-item {
+.otp .option .default-itin .leg .stop-list .stop-item {
   padding-left: 0px;
 }
 
-.otp .option .leg .stop-list .info-item {
+.otp .option .default-itin .leg .stop-list .info-item {
   padding-left: 12px;
 }
 
-/*.otp .option .leg .leg-duration {
-  float: right;
-  font-size: x-small;
-}*/
-
 /* ITINERARY STEP */
 
-.otp .option .step {
-  cursor: pointer;
-  background-color: #eee;
-  margin-bottom: 4px;
-}
-
-.otp .option .step {
+.otp .option .default-itin .step {
   cursor: pointer;
   background-color: #eee;
   margin-bottom: 4px;
@@ -129,15 +118,15 @@
   font-size: x-small;
 }
 
-.otp .option .step:hover {
+.otp .option .default-itin .step:hover {
   background-color: var(--hover-color);
 }
 
-.otp .option .step.active {
+.otp .option .default-itin .step.active {
   background-color: #ccc;
   /*border: 1px dotted;*/
 }
 
-.otp .option .step .step-distance {
+.otp .option .default-itin .step .step-distance {
   float: right;
 }
